requirements.txt
transformers
gradio
langchain
langchain_community
sentence-transformers
faiss-cpu
chardet
